move exec db to common place
