+++ crates/l2/utils/prover/save_state.rs
@@ -378,7 +378,7 @@ pub fn block_number_has_all_proofs(block_number: u64) -> Result<bool, SaveStateE
 mod tests {
     use ethrex_blockchain::Blockchain;
     use ethrex_storage::{EngineType, Store};
-    use ethrex_vm::backends::revm::execution_db::ExecutionDB;
+    use ethrex_vm::db::ExecutionDB;
 
     use super::*;
     use crate::utils::test_data_io;
