+++ crates/vm/backends/levm/db.rs
@@ -2,7 +2,6 @@ use ethrex_common::U256 as CoreU256;
 use ethrex_common::{Address as CoreAddress, H256 as CoreH256};
 use ethrex_levm::db::Database as LevmDatabase;
 
-use crate::backends::revm::execution_db::ExecutionDB;
 use crate::db::StoreWrapper;
 
 impl LevmDatabase for StoreWrapper {
@@ -53,7 +52,7 @@ impl LevmDatabase for StoreWrapper {
     }
 }
 
-impl LevmDatabase for ExecutionDB {
+impl LevmDatabase for crate::db::ExecutionDB {
     fn get_account_info(&self, address: CoreAddress) -> ethrex_levm::AccountInfo {
         let Some(acc_info) = self.accounts.get(&address) else {
             return ethrex_levm::AccountInfo::default();
